Extract version.ts from sandboxTypes.ts
